Cafeteria, functional, lacking db, #BSPQ22E1-46 add work 15m Development

+++ src/main/java/server/data/domain/Drink.java
@@ -4,9 +4,9 @@ public class Drink {
 
     private String name;
     private String type;
-    private float price;
+    private double price;
 
-    public Drink(String name, String type, float price) {
+    public Drink(String name, String type, double price) {
         this.name = name;
         this.type = type;
         this.price = price;
@@ -15,7 +15,7 @@ public class Drink {
     public Drink(){
         this.name = "";
         this.type = "";
-        this.price = 0;
+        this.price = 0.00;
     }
 
     public String getName() {
@@ -32,10 +32,10 @@ public class Drink {
         this.type = type;
     }
 
-    public float getPrice() {
+    public double getPrice() {
         return price;
     }
-    public void setPrice(float price) {
+    public void setPrice(double price) {
         this.price = price;
     }
